Add `GitRun::Repository`
`GitRun.repository` returns a `GitRun::Repository`, which inherits from
`Rugged::Repository` and adds the `#in_revision` method, which switches
to a revision, executes a block and returns to master.

It might be a good idea to turn `.run` into an instance method and run
it on an instance of `GitRun` if we need to set repository paths.